+++ flutter_mxlogger/example/lib/log_page.dart
@@ -31,19 +31,18 @@ class _LogPageState extends State<LogPage> {
     Directory directory = await  getApplicationDocumentsDirectory();
 
 
-
     _mxLogger = await MXLogger.initialize(
         nameSpace: "flutter.mxlogger",
         directory: directory.path,
         storagePolicy: MXStoragePolicyType.yyyy_MM_dd_HH,
-        fileHeader: "version 1.2.5",
+        fileHeader: "version 1.2.8",
         cryptKey: _cryptKey,
         iv: _iv);
 
     _mxLogger.setMaxDiskAge(60*60*24*7);
     _mxLogger.setMaxDiskSize(1024*1024*10);
     _mxLogger.setConsoleEnable(true);
-    _mxLogger.setFileLevel(0);
+    _mxLogger.setLevel(0);
      updateSize();
 
     print("path:${_mxLogger.diskcachePath}");